# search: Deduplicate results w.r.t. repositories