@keegancsmith 2019-10-30
This PR originally had some optimizations to dedup. But I was interested if they are actually necessary. After benchmarking, it seems not. So I dropped the commit from the PR and force pushed.