@keegancsmith 2019-10-30
> This PR originally had some optimizations to dedup. But I was interested if they are actually necessary. After benchmarking, it seems not. So I dropped the commit from the PR and force pushed.

The benchmark was broken! In fact the perf commit did make quite a significant difference, so I've added it back :)

```
name     old time/op  new time/op  delta
Dedup-8   464µs ± 4%   310µs ± 4%  -33.09%  (p=0.008 n=5+5)
```